+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -8,7 +8,7 @@
     * `umElemPtrs :: UMElem c -> Maybe (Set Ptr)`
     * `umElemDRep :: UMElem c -> Maybe (Credential 'Voting (EraCrypto era))`
     * `UView (DRepUView)` constructor and `dRepUView`
-    * `revPtrMap :: UMap c -> Map (Credential 'Staking c) (Set Ptr)`
+    * `invPtrMap :: UMap c -> Map (Credential 'Staking c) (Set Ptr)`
     * `dRepMap :: UMap c -> Map (Credential 'Staking c) (Credential 'Voting c)`
     * synonym `unionL = (∪)`
     * synonym `unionR = (⨃)`
@@ -39,7 +39,7 @@
     * `sumDepositView` to `sumDepositUView`
   * Reimplementations
     * `unionRewAgg` NOTE: It does not require `assert (Map.valid result) result` any more
-      and has been tested for equivalence with the older version with 
+      and has been tested for equivalence with the older version with
       `--qc-max-success=10000 --qc-max-size=1000`. The test is added to `UMapSpec`.
 * Add `certsTxBodyL` to `EraTxBody`
 * Introduce `TxCert` type family and `EraTxCert` type class.
@@ -49,6 +49,12 @@
 * Add new key roles: `CommitteeColdKey` and `CommitteeHotKey`
 * Remove `ConstitutionalDelegCert`. Instead it now lives in `cardano-ledger-shelley` as
   `GenesisDelegCert`
+* Add `StakeCredentials` and two helper functions: `toStakeCredentials` and
+  `domRestrictedStakeCredentials`
+
+### `testlib`
+
+* Add `genValidUMapWithCreds` and `uniformSubset`
 
 ## 1.2.0.0